# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DS Engineosys API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /departments:
    get:
      operationId: getDepartments
      tags: [departments]
      summary: Get all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: Get all employees
      parameters:
        - name: departmentId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Register a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeRequest"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /analytics/product-prediction/{id}:
    get:
      operationId: getProductPrediction
      tags: [analytics]
      summary: Get ML prediction for a product's market performance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPrediction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [ds_engineer, manager, admin]
      required: [name, email, password, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        message:
          type: string
      required: [user, message]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ds_engineer, manager, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        employeeCount:
          type: integer
      required: [id, name, employeeCount]
    CreateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
      required: [name]
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        employeeId:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        designation:
          type: string
        joiningDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, on_leave]
        performanceScore:
          type: number
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required: [id, name, email, employeeId, departmentId, departmentName, designation, joiningDate, status]
    CreateEmployeeRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        employeeId:
          type: string
        departmentId:
          type: integer
        designation:
          type: string
        joiningDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, on_leave]
      required: [name, email, employeeId, departmentId, designation, joiningDate, status]
    UpdateEmployeeRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        departmentId:
          type: integer
        designation:
          type: string
        status:
          type: string
          enum: [active, inactive, on_leave]
...
    ProductPrediction:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        predictedDemand:
          type: string
          enum: [high, medium, low]
        marketLongevityMonths:
          type: integer
        profitScore:
          type: number
        recommendation:
          type: string
          enum: [keep, offer_discount, discontinue, promote]
        suggestedOfferPercentage:
          type: number
          nullable: true
        confidence:
          type: number
        insights:
          type: array
          items:
            type: string
      required: [productId, productName, predictedDemand, marketLongevityMonths, profitScore, recommendation, confidence, insights]
...