# lib/api-zod/src/generated/types/productPrediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DS Engineosys API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductPredictionPredictedDemand } from "./productPredictionPredictedDemand";
import type { ProductPredictionRecommendation } from "./productPredictionRecommendation";

export interface ProductPrediction {
  productId: number;
  productName: string;
  predictedDemand: ProductPredictionPredictedDemand;
  marketLongevityMonths: number;
  profitScore: number;
  recommendation: ProductPredictionRecommendation;
  suggestedOfferPercentage?: number | null;
  confidence: number;
  insights: string[];
}